+++ wedding/src/test/java/project/wedding/cardboard/domain/CardBoardTest.java
@@ -1,4 +1,4 @@
-package project.wedding.domain;
+package project.wedding.cardboard.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,6 +6,9 @@ import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import project.wedding.card.constant.CardStatus;
+import project.wedding.card.domain.Card;
 
 class CardBoardTest {
