+++ wedding/src/test/java/project/wedding/todo/domain/TodoTest.java
@@ -1,5 +1,4 @@
-package project.wedding.domain;
-
+package project.wedding.todo.domain;
 
 import static org.assertj.core.api.Assertions.*;
